\name{pasteAsComment-package}
\alias{pasteAsComment-package}
\title{
  Paste as a comment block or as a 'roxygen' block
}
\description{
  Provides a 'RStudio' addin allowing to paste a selection as a comment block
  or a 'roxygen' block.
}
\details{
  This package does not provide any function, only a 'RStudio' addin.
}
\author{
Stéphane Laurent.

Maintainer: Stéphane Laurent <laurent_step@outlook.fr>
}
\keyword{ package }
